Moved schemas where they belong.
