+++ src/backend/aspen/api/schemas/base.py
@@ -1,6 +1,9 @@
 from pydantic import BaseModel
 
 
-class Base(BaseModel):
+class BaseRequest(BaseModel):
+    pass
+
+class BaseResponse(BaseModel):
     class Config:
         orm_mode = True
