+++ src/backend/aspen/api/schemas/users.py
@@ -1,14 +1,8 @@
+from aspen.api.schemas.base import BaseResponse
 from typing import List
 
-from pydantic import BaseModel
 
-
-class Base(BaseModel):
-    class Config:
-        orm_mode = True
-
-
-class UserBase(Base):
+class UserBase(BaseResponse):
     agreed_to_tos: bool = False
 
 
@@ -16,5 +10,5 @@ class User(UserBase):
     pass
 
 
-class Users(Base):
+class Users(BaseResponse):
     items: List[User]
